fix: Turns out jmdns doesn't work on Android
This change turns off all the zeroconf on the Android target
