+++ mockzilla-management-ui/common/src/commonMain/kotlin/com/apadmi/mockzilla/desktop/di/UseCaseModule.kt
@@ -4,24 +4,28 @@ import com.apadmi.mockzilla.desktop.engine.connection.AdbConnectorService
 import com.apadmi.mockzilla.desktop.engine.connection.AdbConnectorServiceImpl
 import com.apadmi.mockzilla.desktop.engine.connection.DeviceDetectionUseCase
 import com.apadmi.mockzilla.desktop.engine.connection.DeviceDetectionUseCaseImpl
-import com.apadmi.mockzilla.desktop.engine.connection.jmdns.JmdnsWrapper
+import com.apadmi.mockzilla.desktop.engine.connection.ZeroConfSdkWrapper
 import com.apadmi.mockzilla.desktop.engine.device.MetaDataUseCase
 import com.apadmi.mockzilla.desktop.engine.device.MetaDataUseCaseImpl
 import com.apadmi.mockzilla.desktop.engine.device.MonitorLogsUseCase
 import com.apadmi.mockzilla.desktop.engine.device.MonitorLogsUseCaseImpl
 import com.apadmi.mockzilla.lib.config.ZeroConfConfig
+
 import org.koin.core.module.Module
 import org.koin.dsl.module
+
 import java.net.InetAddress
+
+import kotlinx.coroutines.GlobalScope
 
 internal fun useCaseModule(): Module = module {
     single<MetaDataUseCase> { MetaDataUseCaseImpl(get()) }
     single<MonitorLogsUseCase> { MonitorLogsUseCaseImpl(get()) }
     single<AdbConnectorService> { AdbConnectorServiceImpl }
-    single { JmdnsWrapper(ZeroConfConfig.serviceType + ".local.") }
+    single { ZeroConfSdkWrapper(ZeroConfConfig.serviceType + ".local.", GlobalScope) }
     single<DeviceDetectionUseCase> {
         DeviceDetectionUseCaseImpl({ InetAddress.getLocalHost().hostAddress }, get()).also {
-            get<JmdnsWrapper>().setListener(it::onChangedServiceEvent)
+            get<ZeroConfSdkWrapper>().setListener(it::onChangedServiceEvent)
         }
     }
 }
