+++ mockzilla-management-ui/common/src/commonMain/kotlin/com/apadmi/mockzilla/desktop/engine/connection/DeviceDetectionUseCase.kt
@@ -1,6 +1,5 @@
 package com.apadmi.mockzilla.desktop.engine.connection
 
-import com.apadmi.mockzilla.desktop.engine.connection.jmdns.ServiceInfoWrapper
 import com.apadmi.mockzilla.lib.models.MetaData
 import com.apadmi.mockzilla.lib.models.MetaData.Companion.parseMetaData
 import com.apadmi.mockzilla.lib.models.RunTarget
