+++ mockzilla-management-ui/common/src/desktopMain/kotlin/com/apadmi/mockzilla/desktop/jmds/ServiceInfoWrapper.kt
@@ -1,0 +1,17 @@
+package com.apadmi.mockzilla.desktop.jmds
+
+import com.apadmi.mockzilla.desktop.engine.connection.ServiceInfoWrapper
+import javax.jmdns.ServiceInfo
+
+internal fun ServiceInfoWrapper.Companion.create(
+    info: ServiceInfo,
+    hostAddresses: List<String>,
+    state: ServiceInfoWrapper.State
+) = ServiceInfoWrapper(
+    connectionName = info.name,
+    hostAddress = info.hostAddress,
+    hostAddresses = hostAddresses,
+    attributes = info.propertyNames.toList().associateWith { info.getPropertyString(it) },
+    port = info.port,
+    state = state
+)
